Move model.Fragment interfaces to helper
