+++ helper/fragment.go
@@ -1,4 +1,4 @@
-package model
+package helper
 
 import "io"
