+++ raw/wld_write.go
@@ -7,7 +7,7 @@ import (
 	"io"
 
 	"github.com/xackery/encdec"
-	"github.com/xackery/quail/model"
+	"github.com/xackery/quail/helper"
 	"github.com/xackery/quail/raw/rawfrag"
 )
 
@@ -15,7 +15,7 @@ import (
 func (wld *Wld) Write(w io.Writer) error {
 	var err error
 	if wld.Fragments == nil {
-		wld.Fragments = []model.FragmentReadWriter{&rawfrag.WldFragDefault{}}
+		wld.Fragments = []helper.FragmentReadWriter{&rawfrag.WldFragDefault{}}
 	}
 
 	enc := encdec.NewEncoder(w, binary.LittleEndian)
